+++ changelog.md
@@ -1,4 +1,4 @@
-### Next version
+### 0.19.10 (2025-02-15)
 * Upgraded logback from 1.5.6 to 1.5.16
 * Fixed issue in CompetingConsumerSubscriptionModel in which the model could be started automatically if consumption was granted, even though it was explicitly stopped
 * Fixed ConcurrentModificationException's that could occur when doing queries and writing at the same time to an `InMemoryEventStore` instance